+++ examples/angular/src/app/components/form/form.component.ts
@@ -3,6 +3,7 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 
 import {
   RVAccountingProviders,
+  RVOptions,
   RVReportFrequency,
   RVReportTypes,
 } from '@railzai/railz-visualizations';
@@ -21,7 +22,8 @@ import { Filter } from '../../../types/form-submission';
 export class FormComponent implements OnInit {
   @Output() filterEvent = new EventEmitter<Filter>();
   @Output() authEvent = new EventEmitter<string>();
-  @Input() options?: string;
+  @Output() optionEvent = new EventEmitter<RVOptions>();
+  @Input() options?: RVOptions;
 
   authForm = new FormGroup({
     authUrl: new FormControl('', Validators.required),
@@ -44,8 +46,7 @@ export class FormComponent implements OnInit {
       RVReportTypes.BANK_ACCOUNT,
       RVReportTypes.CREDIT_SCORE,
     ].includes(type as RVReportTypes);
-  requiresNoDate = (type: string) =>
-    [RVReportTypes.BANK_ACCOUNT, RVReportTypes.CREDIT_SCORE].includes(type as RVReportTypes);
+  requiresNoDate = (type: string) => [RVReportTypes.BANK_ACCOUNT].includes(type as RVReportTypes);
 
   token = '';
 
@@ -140,6 +141,12 @@ export class FormComponent implements OnInit {
   onFilterSubmit() {
     if (this.filterForm && this.filterForm.valid) {
       this.filterEvent.emit(this.filterForm.value);
+    }
+  }
+
+  onOptionsSubmit(options: RVOptions) {
+    if (options) {
+      this.optionEvent.emit(options);
     }
   }