+++ packages/components/src/components.d.ts
@@ -9,18 +9,19 @@ import {
   RVConfiguration,
   RVFilterAllReportTypes,
   RVFilterBankAccount,
+  RVFilterCreditScore,
   RVFilterFinancialRatio,
-  RVFilterGauge,
-  RVFilterPie,
+  RVFilterIncomeStatementsType,
   RVFilterStatements,
   RVFilterTransactions,
   RVOptions,
   RVOptionsBarStyle,
   RVOptionsPercentageStyle,
+  RVOptionsRatioSparkLineStyle,
   RVPeriodData,
   RVReportTypes,
   RVSelectStyle,
-  RVTooltipStyle,
+  RVTooltipIndicatorStyle,
 } from './types';
 export namespace Components {
   interface RailzBankAccounts {
@@ -32,6 +33,20 @@ export namespace Components {
      * Filter information to query the backend APIs
      */
     filter: RVFilterBankAccount;
+    /**
+     * For whitelabeling styling
+     */
+    options: RVOptions;
+  }
+  interface RailzCreditScore {
+    /**
+     * Configuration information like authentication configuration
+     */
+    configuration: RVConfiguration;
+    /**
+     * Filter information to query the backend APIs
+     */
+    filter: RVFilterCreditScore;
     /**
      * For whitelabeling styling
      */
@@ -51,7 +66,7 @@ export namespace Components {
      */
     statusCode?: number;
     /**
-     * Color of the image text
+     * Style of the image text
      */
     textStyle?: { [key: string]: any };
     /**
@@ -73,7 +88,7 @@ export namespace Components {
      */
     options: RVOptions;
   }
-  interface RailzGaugeChart {
+  interface RailzIncomeStatements {
     /**
      * Configuration information like authentication configuration
      */
@@ -81,7 +96,7 @@ export namespace Components {
     /**
      * Filter information to query the backend APIs
      */
-    filter: RVFilterGauge;
+    filter: RVFilterIncomeStatementsType;
     /**
      * For whitelabeling styling
      */
@@ -115,20 +130,6 @@ export namespace Components {
      */
     percentage: number;
     percentageStyle?: RVOptionsPercentageStyle;
-  }
-  interface RailzPieChart {
-    /**
-     * Configuration information like authentication configuration
-     */
-    configuration: RVConfiguration;
-    /**
-     * Filter information to query the backend APIs
-     */
-    filter: RVFilterPie;
-    /**
-     * For whitelabeling styling
-     */
-    options: RVOptions;
   }
   interface RailzProgressBar {
     /**
@@ -167,7 +168,7 @@ export namespace Components {
      * Data to display for sparkline
      */
     data: Array<RVPeriodData>;
-    sparkLineStyle?: { [key: string]: any };
+    sparkLineStyle?: RVOptionsRatioSparkLineStyle;
   }
   interface RailzStatementsChart {
     /**
@@ -191,7 +192,7 @@ export namespace Components {
     /**
      * Position of the Tooltip text when hovered
      */
-    tooltipStyle?: RVTooltipStyle;
+    tooltipStyle?: RVTooltipIndicatorStyle;
     tooltipText: string;
   }
   interface RailzTransactionsControl {
@@ -229,6 +230,11 @@ declare global {
     prototype: HTMLRailzBankAccountsElement;
     new (): HTMLRailzBankAccountsElement;
   };
+  interface HTMLRailzCreditScoreElement extends Components.RailzCreditScore, HTMLStencilElement {}
+  var HTMLRailzCreditScoreElement: {
+    prototype: HTMLRailzCreditScoreElement;
+    new (): HTMLRailzCreditScoreElement;
+  };
   interface HTMLRailzErrorImageElement extends Components.RailzErrorImage, HTMLStencilElement {}
   var HTMLRailzErrorImageElement: {
     prototype: HTMLRailzErrorImageElement;
@@ -241,10 +247,12 @@ declare global {
     prototype: HTMLRailzFinancialRatiosElement;
     new (): HTMLRailzFinancialRatiosElement;
   };
-  interface HTMLRailzGaugeChartElement extends Components.RailzGaugeChart, HTMLStencilElement {}
-  var HTMLRailzGaugeChartElement: {
-    prototype: HTMLRailzGaugeChartElement;
-    new (): HTMLRailzGaugeChartElement;
+  interface HTMLRailzIncomeStatementsElement
+    extends Components.RailzIncomeStatements,
+      HTMLStencilElement {}
+  var HTMLRailzIncomeStatementsElement: {
+    prototype: HTMLRailzIncomeStatementsElement;
+    new (): HTMLRailzIncomeStatementsElement;
   };
   interface HTMLRailzLoadingElement extends Components.RailzLoading, HTMLStencilElement {}
   var HTMLRailzLoadingElement: {
@@ -255,11 +263,6 @@ declare global {
   var HTMLRailzPercentageElement: {
     prototype: HTMLRailzPercentageElement;
     new (): HTMLRailzPercentageElement;
-  };
-  interface HTMLRailzPieChartElement extends Components.RailzPieChart, HTMLStencilElement {}
-  var HTMLRailzPieChartElement: {
-    prototype: HTMLRailzPieChartElement;
-    new (): HTMLRailzPieChartElement;
   };
   interface HTMLRailzProgressBarElement extends Components.RailzProgressBar, HTMLStencilElement {}
   var HTMLRailzProgressBarElement: {
@@ -306,12 +309,12 @@ declare global {
   };
   interface HTMLElementTagNameMap {
     'railz-bank-accounts': HTMLRailzBankAccountsElement;
+    'railz-credit-score': HTMLRailzCreditScoreElement;
     'railz-error-image': HTMLRailzErrorImageElement;
     'railz-financial-ratios': HTMLRailzFinancialRatiosElement;
-    'railz-gauge-chart': HTMLRailzGaugeChartElement;
+    'railz-income-statements': HTMLRailzIncomeStatementsElement;
     'railz-loading': HTMLRailzLoadingElement;
     'railz-percentage': HTMLRailzPercentageElement;
-    'railz-pie-chart': HTMLRailzPieChartElement;
     'railz-progress-bar': HTMLRailzProgressBarElement;
     'railz-select': HTMLRailzSelectElement;
     'railz-sparkline-chart': HTMLRailzSparklineChartElement;
@@ -336,6 +339,20 @@ declare namespace LocalJSX {
      */
     options?: RVOptions;
   }
+  interface RailzCreditScore {
+    /**
+     * Configuration information like authentication configuration
+     */
+    configuration: RVConfiguration;
+    /**
+     * Filter information to query the backend APIs
+     */
+    filter: RVFilterCreditScore;
+    /**
+     * For whitelabeling styling
+     */
+    options?: RVOptions;
+  }
   interface RailzErrorImage {
     /**
      * Fill color of the svg image representing a status code
@@ -350,7 +367,7 @@ declare namespace LocalJSX {
      */
     statusCode?: number;
     /**
-     * Color of the image text
+     * Style of the image text
      */
     textStyle?: { [key: string]: any };
     /**
@@ -372,7 +389,7 @@ declare namespace LocalJSX {
      */
     options?: RVOptions;
   }
-  interface RailzGaugeChart {
+  interface RailzIncomeStatements {
     /**
      * Configuration information like authentication configuration
      */
@@ -380,7 +397,7 @@ declare namespace LocalJSX {
     /**
      * Filter information to query the backend APIs
      */
-    filter: RVFilterGauge;
+    filter: RVFilterIncomeStatementsType;
     /**
      * For whitelabeling styling
      */
@@ -414,20 +431,6 @@ declare namespace LocalJSX {
      */
     percentage?: number;
     percentageStyle?: RVOptionsPercentageStyle;
-  }
-  interface RailzPieChart {
-    /**
-     * Configuration information like authentication configuration
-     */
-    configuration: RVConfiguration;
-    /**
-     * Filter information to query the backend APIs
-     */
-    filter: RVFilterPie;
-    /**
-     * For whitelabeling styling
-     */
-    options?: RVOptions;
   }
   interface RailzProgressBar {
     /**
@@ -467,7 +470,7 @@ declare namespace LocalJSX {
      * Data to display for sparkline
      */
     data: Array<RVPeriodData>;
-    sparkLineStyle?: { [key: string]: any };
+    sparkLineStyle?: RVOptionsRatioSparkLineStyle;
   }
   interface RailzStatementsChart {
     /**
@@ -491,7 +494,7 @@ declare namespace LocalJSX {
     /**
      * Position of the Tooltip text when hovered
      */
-    tooltipStyle?: RVTooltipStyle;
+    tooltipStyle?: RVTooltipIndicatorStyle;
     tooltipText: string;
   }
   interface RailzTransactionsControl {
@@ -524,12 +527,12 @@ declare namespace LocalJSX {
   }
   interface IntrinsicElements {
     'railz-bank-accounts': RailzBankAccounts;
+    'railz-credit-score': RailzCreditScore;
     'railz-error-image': RailzErrorImage;
     'railz-financial-ratios': RailzFinancialRatios;
-    'railz-gauge-chart': RailzGaugeChart;
+    'railz-income-statements': RailzIncomeStatements;
     'railz-loading': RailzLoading;
     'railz-percentage': RailzPercentage;
-    'railz-pie-chart': RailzPieChart;
     'railz-progress-bar': RailzProgressBar;
     'railz-select': RailzSelect;
     'railz-sparkline-chart': RailzSparklineChart;
@@ -545,16 +548,17 @@ declare module '@stencil/core' {
     interface IntrinsicElements {
       'railz-bank-accounts': LocalJSX.RailzBankAccounts &
         JSXBase.HTMLAttributes<HTMLRailzBankAccountsElement>;
+      'railz-credit-score': LocalJSX.RailzCreditScore &
+        JSXBase.HTMLAttributes<HTMLRailzCreditScoreElement>;
       'railz-error-image': LocalJSX.RailzErrorImage &
         JSXBase.HTMLAttributes<HTMLRailzErrorImageElement>;
       'railz-financial-ratios': LocalJSX.RailzFinancialRatios &
         JSXBase.HTMLAttributes<HTMLRailzFinancialRatiosElement>;
-      'railz-gauge-chart': LocalJSX.RailzGaugeChart &
-        JSXBase.HTMLAttributes<HTMLRailzGaugeChartElement>;
+      'railz-income-statements': LocalJSX.RailzIncomeStatements &
+        JSXBase.HTMLAttributes<HTMLRailzIncomeStatementsElement>;
       'railz-loading': LocalJSX.RailzLoading & JSXBase.HTMLAttributes<HTMLRailzLoadingElement>;
       'railz-percentage': LocalJSX.RailzPercentage &
         JSXBase.HTMLAttributes<HTMLRailzPercentageElement>;
-      'railz-pie-chart': LocalJSX.RailzPieChart & JSXBase.HTMLAttributes<HTMLRailzPieChartElement>;
       'railz-progress-bar': LocalJSX.RailzProgressBar &
         JSXBase.HTMLAttributes<HTMLRailzProgressBarElement>;
       'railz-select': LocalJSX.RailzSelect & JSXBase.HTMLAttributes<HTMLRailzSelectElement>;