+++ packages/components/src/elements/progress-bar/progress-bar.tsx
@@ -53,16 +53,16 @@ export class ProgressBar {
       return <span></span>;
     }
     return (
-      <div class="railz-progress-bar-div" style={this.options?.divStyle}>
-        {this._options.noTitle ? null : (
+      <div class="rv-progress-bar-div" style={this.options?.divStyle}>
+        {this._options.hideLabels ? null : (
           <p
-            class="railz-progress-bar-total-unpaid"
+            class="rv-progress-bar-total-unpaid"
             style={this._options?.titleStyle}
             aria-label={`${Translations.RV_TOTAL_UNPAID} ${formatNumber(this.unpaidAmount)}`}
           >
             {Translations.RV_TOTAL_UNPAID} {this.titleMapping[this.reportType] || ''}{' '}
             <span
-              class="railz-progress-bar-total-unpaid-value"
+              class="rv-progress-bar-total-unpaid-value"
               style={this._options?.titleValueStyle}
               aria-hidden={true}
             >
@@ -71,19 +71,19 @@ export class ProgressBar {
           </p>
         )}
 
-        <div class="railz-progress-bar-values">
-          {this._options.noTitle ? null : (
-            <div class="railz-progress-bar-values-div">
+        <div class="rv-progress-bar-values">
+          {this._options.hideLabels ? null : (
+            <div class="rv-progress-bar-values-div">
               <div>
                 <p
-                  class="railz-progress-bar-label"
+                  class="rv-progress-bar-label"
                   style={this._options?.subTitle1Style}
                   aria-label={`${Translations.RV_PAID} ${formatNumber(this.paidAmount)}`}
                 >
                   {Translations.RV_PAID}
                 </p>
                 <p
-                  class="railz-progress-bar-value"
+                  class="rv-progress-bar-value"
                   style={this._options?.subTitleValue1Style}
                   aria-hidden={true}
                 >
@@ -92,14 +92,14 @@ export class ProgressBar {
               </div>
               <div>
                 <p
-                  class="railz-progress-bar-label railz-progress-bar-overdue"
+                  class="rv-progress-bar-label rv-progress-bar-overdue"
                   style={this._options?.subTitle2Style}
                   aria-label={`${Translations.RV_OVERDUE} ${formatNumber(this.overdueAmount) || 0}`}
                 >
                   {Translations.RV_OVERDUE}
                 </p>
                 <p
-                  class="railz-progress-bar-value railz-progress-bar-overdue"
+                  class="rv-progress-bar-value rv-progress-bar-overdue"
                   style={this._options?.subTitleValue2Style}
                   aria-hidden={true}
                 >
@@ -108,12 +108,10 @@ export class ProgressBar {
               </div>
             </div>
           )}
-          <div class="railz-progress-bar" style={this._options?.barStyle}>
+          <div class="rv-progress-bar" style={this._options?.barStyle}>
             <span
               style={{
-                width: this._options?.customWidth
-                  ? this._options?.customWidth
-                  : `${this.normalize(this.paidAmount, this.unpaidAmount + this.paidAmount)}%`,
+                width: `${this.normalize(this.paidAmount, this.unpaidAmount + this.paidAmount)}%`,
                 ...this._options?.progressStyle,
               }}
             />