+++ packages/components-angular/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.7 (2022-07-07)
+
+**Note:** Version bump only for package @railzai/railz-visualizations-angular
+
+
+
+
+
+## 0.0.7-alpha.0 (2022-07-07)
+
+**Note:** Version bump only for package @railzai/railz-visualizations-angular
+
+
+
+
+
+## 0.0.6 (2022-07-07)
+
+
+### Features
+
+*  Include more customization options for components ([60b8d33](https://github.com/railz-ai/railz-visualizations/commit/60b8d3387b06007dd5f0ebeece8a9e9cc79c7864))
+
+
+
+
+
+## 0.0.5 (2022-07-07)
+
+
+### Features
+
+* [skip actions] Update namings and documentation for components ([#100](https://github.com/railz-ai/railz-visualizations/issues/100)) ([a63e25c](https://github.com/railz-ai/railz-visualizations/commit/a63e25c274e2c4ca6217aaefb9005538826ae38e))
+
+
+
+
+
 ## 0.0.5-alpha.12 (2022-07-05)